+++ Emote.cpp
@@ -15,13 +15,13 @@ Emote::Emote(const sf::Texture& texture)
     m_sprite.setTexture(*m_texture);
     for( int i = 0; i < m_frame; i++ )
     {
-        m_anim.push_back(sf::IntRect(i*m_size.x, 0, m_size.x, m_size.y));
+        m_anim.push_back(sf::IntRect(static_cast<int>(i*m_size.x), 0, static_cast<int>(m_size.x), static_cast<int>(m_size.y) ));
     }
     
     m_sprite.setTextureRect(m_anim[0]);
     //m_sprite.setPosition(newPos);
-    m_sprite.setScale(0.6,0.6);
-    m_sprite.setOrigin(sf::Vector2f(m_size.x/2, m_size.y));
+    m_sprite.setScale(0.6f,0.6f);
+    m_sprite.setOrigin(sf::Vector2f(m_size.x/2.f, m_size.y));
 }
 
 Emote::~Emote()
